Extract other robot poses to separate map panel
